Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-307-status-code.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var tests =
[
 { responseHeaders: {'Status': '307'} },
 { responseHeaders: {'Status': '307', 'Cache-control': 'max-age=0'} },
 { responseHeaders: {'Status': '307', 'Cache-control': 'max-age=100'} },
 { responseHeaders: {'Status': '307', 'Expires': 'now(0)'} },
 { responseHeaders: {'Status': '307', 'Expires': 'now(100)'} },
 { responseHeaders: {'Status': '307', 'Last-Modified': 'now(-1000)'} },
];

description("Test that responses with HTTP status code 307 which are not cacheable by default");

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
