LayoutTests/http/tests/cache/disk-cache/disk-cache-307-status-code-expected.txt

Test that responses with HTTP status code 307 which are not cacheable by default

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 6 tests

response headers: {"Status":"307"}
response source: Network

response headers: {"Status":"307","Cache-control":"max-age=0"}
response source: Network

response headers: {"Status":"307","Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Status":"307","Expires":"now(0)"}
response source: Network

response headers: {"Status":"307","Expires":"now(100)"}
response source: Disk cache

response headers: {"Status":"307","Last-Modified":"now(-1000)"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
